course3/msa/chapter3/dashboard-app/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
jinja2==3.0
aiofiles==22.1.0
pydantic==1.10.4
sqlalchemy==1.4.46
pymysql==1.0.2
kafka-python==2.0.2
# confluent-kafka==1.9.2
websockets==10.4
